github-ci: check for new authors
Add a GitHub CI job that checks the commits in a pull requests for new
authors that have not been seen before. If a new author is found, the
job will error out with the unknown authors listed as error annotations.

Note the list of previous authors is taken from the master branch, so
there is a chance an author might not be detected as an existing author.

The GitHub action for checking out code is purposefully not used as it
has shown to provided incomplete author information in previous tests.