output/tx: attempt to fix extra applayer pkt log
Currently, we have what looks like a bug where after a packet drop the
engine will log the app-layer event associated with the packet just
dropped.

This seems connected with the Tx logging. But this commit's approach
leads to wrong count of some alerts with threshold config and, what's
worse, in triggering of the debug validation for calling stats_incr with
id 0...