exceptions/midstream: parse midstream policy alone
As the midstream exception policy has its own specific scenarios, have a
dedicated function to parse and process its config values, and check for
midstream enabled when needed.

What is 'auto' will differ if midstream is enabled, for
midstream-policy. Make sure we check and respect that, too.

TODO:

- review all sv tests
- refine it. It's ugly af
- bring missing TODOs and questions from other commit (different branch)

Related to
Bug #5825