detect-file-data: inspect http body if in progress
Detect the HTTP response body even if in the state of processing the
HTTP body. This is required for the filestore to save the file,
otherwise inspection is not done until after the file has been pruned.

Ideally this would only be done when the filestore is enabled, but I
don't see a non-intrusive way to determine that here.

Ticket: #5868