util/mime: don't fail on B64_BLOCK
While it is checked in some callers to only send data that is over
B64_BLOCK, an edge case such as multiple spaces or invalid chars in the
last some blocks of data may need stripping of those invalid chars.
Remainder handler is perfectly capable of handling data up to B64_BLOCK
in length. And, that's what the limit should be.

Ticket 6207