conf: add ConfGetChildValueString fn
Although the fn ConfGetChildValue already retrieves string, if a node
exists but does not have a value, its a valid node as that node might be
parent to some children. However, this is invalid if the node in
question was expected to be a string. In order to deal with this, follow
the ways currently integers are dealt with and add a fn
ConfGetChildValueString and ConfGetChildValueStringWithDefault where the
value is expected to be a string and is known to not be a parent node.

Related to Redmine ticket 6249