packetpool: Lock return stack before checking if empty
Previously the return stack queue was checked before obtaining the
lock. This leaves a small window where the FlowManager thread is
waiting for the queue lock and another thread is populating the queue.
The FlowManager thread would miss the signal as it had not started
waiting for it. The FlowManager thread would be stuck forever.

The overall result of this is that there are no available packet
buffers to be used for forcing closure of timed out flows, we run
out of flows and traffic is blocked.

Bug: #6292